TransIter: add function for conversion to a TransPrioQueue
This crate is centered around the `TransIter` type. The `IntoTransIter`
and `AutoTransIter` traits thus deal in terms of that type, and we want
it to stay that way. Still, we want to enable creating `TransPrioQueue`s
in an easy way.

The conversion introduced by this change provides an easy way out.