tests: use depth-based sectioning in node_order_breadth_first
With our previous strategy for this test, we ended up spending a lot of
time doing a costly search of a sequence of ids in a presumably long
list of node ids. This made testing rather painful, with tests sometimes
taking up nearly 10min with `quickcheck`'s defaults on my (arguably
"old") machine.

In the past, we already tried to at least keep the haystack for our
improvised search small by only considering ids after a certain point.
This change migrates to a different approach for reducing the search
space: in breadth-first mode, we'll get the items sorted by depth.
Hence, we can divide the sequence of node ids yielded by our `TransIter`
into sections corresponding to a given depth and restrict the search to
the section we expect the sequences in. In each section, we also apply
our old strategy to reduce the search space further.

With `quickcheck`'s defaults, these changes easily reduces the time a
test-run needs to less than a third. It's still by far the most costly
test we do, but now it's in a regime where it's not perceived as that
much of a pain to run it.

We also take the opportunity to use `Iterator::eq` rather than
constructing it by hand using `Iterator::zip`.